+++ h7code/Core/Src/measure/my_filter.h
@@ -3,5 +3,5 @@
 void MyFilterInit();
 
 
-int32_t SosFilterProcess_x4_0(int32_t newX);
-int32_t SosFilterProcess_x4_1(int32_t newX);
+float SosFilterProcess_x4_0(float newX);
+float SosFilterProcess_x4_1(float newX);
